Added .toml opt, fixed default saving and stuff
