+++ Project/Cargo.toml
@@ -16,3 +16,8 @@ gif = "0.12.0"
 native-dialog = "0.6.4"
 arboard = "3.2.0"
 
+[profile.dev]
+opt-level=2
+
+[profile.release]
+opt-level=3
